tests: Refactor runner tests extract test utils into recipe.rs
